requirements.txt
flask
flask-cors
requests
Markdown
flask-wtf
wtforms
waitress
sendgrid
python-dotenv